Convert ExportOptionsContext from context to state
