+++ packages/lesmis/src/components/DashboardExportModal/components/OptionsBar.js
@@ -9,16 +9,15 @@ import FormControlLabel from '@material-ui/core/FormControlLabel';
 import PropTypes from 'prop-types';
 import { FlexEnd } from '@tupaia/ui-components';
 import Pagination from '@material-ui/lab/Pagination';
-import { useExportOptions } from '../context/ExportOptionsContext';
 import { I18n } from '../../../utils';
 
-export const OptionsBar = ({ totalPage, page, setPage, isDisabled }) => {
+export const OptionsBar = ({ totalPage, page, setPage, isDisabled, useExportOptions }) => {
   const {
     exportWithLabels,
     toggleExportWithLabels,
     exportWithTable,
     toggleExportWithTable,
-  } = useExportOptions();
+  } = useExportOptions;
 
   return (
     <FlexEnd>
@@ -60,4 +59,5 @@ OptionsBar.propTypes = {
   totalPage: PropTypes.number.isRequired,
   isDisabled: PropTypes.bool.isRequired,
   setPage: PropTypes.func.isRequired,
+  useExportOptions: PropTypes.object.isRequired,
 };
